refactor: move AttesterIdentity into identity folder
