+++ src/actor/Claimer.spec.ts
@@ -2,7 +2,7 @@ import Bool from '@polkadot/types/primitive/Bool'
 import AccountId from '@polkadot/types/primitive/Generic/AccountId'
 import { Tuple, Option } from '@polkadot/types/codec'
 import { Text } from '@polkadot/types'
-import AttesterIdentity from '../attesteridentity/AttesterIdentity'
+import AttesterIdentity from '../identity/AttesterIdentity'
 import Identity from '../identity/Identity'
 import constants from '../test/constants'
 import { Attester, Claimer, IClaim, Verifier, CombinedPresentation } from '..'
