+++ src/actor/Verifier.spec.ts
@@ -4,7 +4,7 @@ import { Tuple, Option } from '@polkadot/types/codec'
 import { Text } from '@polkadot/types'
 import { Verifier, Attester, Claimer, CombinedPresentation } from '..'
 import { MessageBodyType } from '../messaging/Message'
-import AttesterIdentity from '../attesteridentity/AttesterIdentity'
+import AttesterIdentity from '../identity/AttesterIdentity'
 import Identity from '../identity/Identity'
 import IClaim from '../types/Claim'
 import constants from '../test/constants'
