+++ src/actor/Verifier.ts
@@ -2,7 +2,7 @@ import * as gabi from '@kiltprotocol/portablegabi'
 import IPublicIdentity from '../types/PublicIdentity'
 import Message, { MessageBodyType, IMessage } from '../messaging/Message'
 import AttestedClaim from '../attestedclaim/AttestedClaim'
-import PublicAttesterIdentity from '../attesteridentity/PublicAttesterIdentity'
+import PublicAttesterIdentity from '../identity/PublicAttesterIdentity'
 import Identity from '../identity/Identity'
 import { factory as LoggerFactory } from '../config/ConfigLog'
 import CType from '../ctype/CType'
