+++ src/identity/AttesterIdentity.ts
@@ -9,7 +9,7 @@ import * as u8aUtil from '@polkadot/util/u8a'
 import * as gabi from '@kiltprotocol/portablegabi'
 import { KeyringPair } from '@polkadot/keyring/types'
 import { IRevocationHandle } from '../types/Attestation'
-import Identity from '../identity/Identity'
+import Identity from './Identity'
 import { MessageBodyType, IInitiateAttestation } from '../messaging/Message'
 import IRequestForAttestation from '../types/RequestForAttestation'
 import PublicAttesterIdentity from './PublicAttesterIdentity'
