+++ src/identity/index.ts
@@ -5,6 +5,14 @@
 
 import Identity from './Identity'
 import PublicIdentity, { IURLResolver } from './PublicIdentity'
+import AttesterIdentity from './AttesterIdentity'
+import PublicAttesterIdentity from './PublicAttesterIdentity'
 
-export { Identity, PublicIdentity, IURLResolver }
+export {
+  Identity,
+  AttesterIdentity,
+  PublicIdentity,
+  PublicAttesterIdentity,
+  IURLResolver,
+}
 export default Identity
